Fix log4j2 for Quickstart.
In QuickStart, we start all Pinot components in a single JVM. And the existing
log4j2 config for it directs package specific logs to different components.
This is broken because packages can be shared across components (eg pinot-core).

During debugging recent problems reported in OSS, we found that realtime
consumption would stop in Pinot, but no errors/exceptions would appear in
the logs. This was due to the fact that the exceptions were thrown by packages
that were not part of any component's log, and were missed.

In this PR we address the problem as:
- We create a single log file with log level `info` for quick-start, as it
  starts all components in single JVM.
- We set the log level to error for console.