Fix CAST transform function for chained transforms (#6941)
Currently the `CAST` transform function relies on the caller to call the correct API to perform the cast. When the `CAST` is chained with other transform functions such as `ADD`, the next transform function might ignore the output type, and ignore the `CAST`. E.g. `CAST(doubleCol AS INT) + 5` will return the same result as `doubleCol + 5`, which is unexpected.
This PR fixes the bug by checking the result type and always cast the values.
Also fix a bug for `TIMESTAMP` cast (missing `break;`)