Enabling compatibility tests in the script (#6959)
The actual run of compat tests were disahled in the script. This PR
enables the compat tests.

Fixed the verifier script to work with relative as well as absolute
paths for working directory.

Other improvements:
We now preserve the log files of all processes we start.

Changed all pid files to be in one dir so it is easy to stop the cluster

Changed the mechanism to check if some service is up. This part needs more
improvement.

As yet, the tests fail after server upgrade