Improvements in compat-tester
Enable users to provide tag names as an alternative to actual git hash.

Made -w mandatory and -o and -n as optional arguments. If -n is
not specified, the current revision is taken to be the newer one.

If -o is also not specfied then the previous commit is taken to be
the older one.

Moved the error messages in compCheck so that they show up after the cluster
is shutdown.