Change default value for `max.parallel.segment.builds`.
We have noticed a lot of instances where realtime consumption runs into issues that trace
back to OOMs related to segment builds. One major contributing factor for this is the
maximum number of parallel segment builds allowed, which defaults to 0, which essentially
means upto number of partitions. More often than not, this is not a good default, especially
for folks who are experiment with Pinot and may have large number of partitions to consume from.

While there's no one value that can satisfy all use cases, the intent here is to limit on a
more reasonable size as opposed to unlimited to number of partitions.