Add Query Options for GroupBy In-Segment Trim  (#7052)
Adds query options for groupBy in-segment trim. Two options are added: segmentMinTrimSize and segmentEnableTrim.

If segmentMinTrimSize is set to be positive in the query option, it will override the server config and trigger the trim based on the given number. If segmentEnableTrim is set to be true in the query option, it will trigger the trim based on the server config trim size if configured, or default trim size (5000).

Example:
minTrimSize in server config: 0
Query: "SELECT metric_0, max(metric_1) FROM testTable GROUP BY metric_0 ORDER BY max(metric_1) DESC LIMIT 50 OPTION(segmentMinTrimSize=1000)")
=> max(5*limit, 1000): 1000
=> so expectedSize from each segment: 1000