Adding support for config files in compat tests
Pinot installation administrators can now specify configuration files to run the
tests as per their environment.

Changed the java files to get the ports as declared in the config files.
Cleaned up the scripts to use variables for ports.
Updated the scripts to compile with java 8 (so that it can be run on desktops
with one java environment for older and newer release)