Add system info on instance config for servers.
There are cases where we want to be able to system info of
instances in the cluster, for example, we would like to know how many
cores, RAM, heap etc do each server for a table in the cluster have.

In this PR, we add this information as a `mapFields` property in the
instance config. Storing this information in ZK can provide a global view
of the cluster, as opposed to each instance individually knowing about itself.
```
  "mapFields": {
    "SYSTEM_RESOURCE_INFO": {
      "numCores": "16",
      "totalMemoryMB": "32768",
      "maxHeapSizeMB": "8192"
    }
  }
```