Fix errors happening due to parallel builds
Maven cache interferes when we build two trees in parallel,
causing (for example) pinot-spi of one tree to be linked
with that of the other tree (since they are of the same version)

Changed the build to use different version ID for old and new builds
so that they don't pull in each other's jars. Current build always
uses the same ID as the current tree so that any changes made
can be tested easily without backing out pom files.

A different build ID each time can cause maven cache to explode if we
run the compile multiple times, so changed the build to use a dedicated
cache that is deleted after builds.

Also added code to cleanup the controller and server's data
directories before starting compat checker.