Introduce OR Predicate Execution On Star Tree Index (#7184)
Allows OR predicates to be executed on Star Tree indices. We only allow a single dimension to
be present in the OR predicate in order to be qualified for execution on Star Tree node. This is to ensure
that we do not double count pre aggregated documents while traversing the tree for two dimensions.