Add github action workflow for ad-hoc compatibility test (#7296)
We currently run the compatibility regression test for every git commit.
On the other hand, we currently don't support to submit the test on
ad-hoc basis. This PR adds the way to submit the test and it also
allows the user to submit 2 inputs for old and new commits. This will
be used as part of Apache release process (comparing 2 releases).

In order to run the test, we can use the Github Action UI that will be
created after this PR.
https://github.com/apache/pinot/actions/workflows/pinot_compatibility_tests.yml