Make the SQL Editor box expandable
The default size of the SQL Editor box is such that it only shows
about 5 lines of text. The SQL Editor uses [react-codemirror2](https://github.com/scniro/react-codemirror2) which recommends
using [re-resizable](https://github.com/bokuweb/re-resizable) to make the editor field resizable.

This patch adds a new NPM library dependency on `re-resizable` and adds
a react component to make the field expandable. As a way to make the change minimal,
the patch only allows the field to be vertically expandable and keeps the default
size the same (height is 100px).