Enhance TableCache to support schema name different from table name (#7525)
Enhance the TableCache to support the case where table schema name is not the same as the raw table name.
We recommend configuring the schema name the same as the raw table name, but it is not enforced. This PR fixes the issue of not able to find the schema from TableCache when its name is not the same as the raw table name. Query optimizers rely on the schema from the TableCache to optimize the query.