In TableRebalancer, allow increasing from 1 replica without downtime (#7532)
Currently when rebalancing a table without downtime, the min available replicas check is performed on the min replica of current and target assignment. This will prevent user from increasing the replica of the table from 1 without downtime. This PR fixes that by only rely on target assignment to decide the min available replicas.