Fix SegmentGenerationWithTimeColumnTest (#7537)
This test would fail when timestamps were generated in the first or last day and the simple date mode was used. There isn't enough information to get a number of milliseconds in UTC. We don't know which timezone the simple date is in, but also don't know that there isn't a timezone in the format string either, so the second change ensures that we don't test within a day of the minimum or maximum allowed number of milliseconds. This has no functional impact, it only relates to interval validation and Pinot wasn't live in 1971, I don't want to make a prediction about 2071.

I added a couple of tests to fix the seed for degenerate cases.