map bitmaps through a bounded window to avoid excessive disk pressure… (#7535)
The assumption behind BitmapInvertedIndexWriter is that there is a lot of disk space available, and that it's ok to overmap and truncate in order to avoid syscall overhead, but this results in high disk usage. Allocating file space for the serialised inverted indexes as it is needed feels like a good usability tradeoff.
The strategy builds on some knowledge about how the bitmaps get serialised: a single bitmap never needs more than 256MB, so we never map more than 256MB at a time. Bitmaps should be fairly sparse most of the time, so we guess that we need numBitmaps * 128KB but if this estimate proves to be too small, the pessimistic 256MB is used for the next section.