[Refactor] Remove the index access within the PlanNode constructor (#7550)
Currently when creating the PlanNodes for the query, there is some complex logic which also accesses the index. The plan creation is done on a single thread within a for loop, so this step can take long when there are lots of segments.
This PR removes the complex logic, and make PlanNode creation very light-weight. The logic of choosing the physical execution plan (Operator) is moved to the PlanNode.run().