Enhance TableRebalancer to offload the segments from most loaded instances first (#7574)
When TableRebalancer rebalances a table without downtime, it goes over multiple steps to reach the target assignment. In order to keep the available replicas, the algorithm picks the first instance in alphabetical order to move the segment. This can cause the instances with smaller name to be overloaded during the intermediate assignments.
This PR enhances the algorithm to track the number of segments to be offloaded for each instance between the current assignment and the target assignment, and always offload the segment from the instances with most segments to be offloaded. With the enhancement, it is guaranteed that during the rebalance, no segment will be moved to the instances to be offloaded.