introduce `LZ4_WITH_LENGTH` chunk compression type (#7655)
All other chunk compression formats can already report their decompressed lengths in O(1) time, but this is missing from the LZ4 specification. The LZ4 library has a compressor which adds the size as a prefix, but this is incompatible with the compression mode currently in use, which requires the introduction of a new Pinot compression mode. Moreover, this new compression mode cannot be relied upon for buffer sizing when reading v1, v2, or v3 of the forward index chunk format because there will always be LZ4 chunks out there with these versions and it will never be possible to determine their chunk sizes, so there is no benefit in using this for newly created v3 chunks.

This PR paves the way to introduce a new chunk format where all chunks are self-describing and reader buffers can be sized from compression metadata. Systems which choose LZ4 compression will be automatically upgraded to LZ4_WITH_LENGTH iff using a new chunk format about to be introduced (but the v3 writer is unaffected). Users of the v3 chunk format can configure LZ4_WITH_LENGTH if they want to, but since it can't be relied on by the reader, this doesn't offer any benefits. LZ4 compressor and decompressor constructors are made package private to enforce (encourage) obtaining them through the factory which enforces the upgrade rule when upgrade is permitted by the caller.