When reloading all segments for a table, do not fail on one bad segment (#7675)
When reloading all segments for a table, currently one bad segment can fail the reload, and leave the rest of the segments not reloaded. This PR fixes that by logging the error, caching the exception, and continue reloading the other segments. In the end, the method throws an exception with the failed segments and the cached exception.