Add forceCleanup option for 'startReplaceSegments' API
The current API will not attempt to proactively clean
up the lineage entry and throw the exception. When the
offline flow fails in the middle of the replacement
protocol, the second attempt will directly fail due to
the duplicate entry for 'segmentsFrom'. In this case,
having 'forceCleanup' option will help us to move forward.

If the flag is on, we proactively revert the previous
lineage entry and delete segments from 'segmentsTo'.