Fix merge delay metric when there is empty bucket (#7761)
when we have the empty buckets (no more data push
for the table). However, we still have some edge
cases not covered. We found the issue where the
table has some empty buckets and new fresh data
that is not eligible for merge due to buffer time.
In this case, we still fall back to the original
algorithm and report the wrong delay. This PR
addresses such case.