Fix the broker routing when segment is deleted (#7817)
# Issue
When a segment is deleted, or moved from one server to another server, it follows these steps:
1. Ideal state is modified by the controller
2. Controller sends messages to servers to offload and delete the segment
3. Servers deletes the segment and update their current state
4. Controller gathers current states of the servers and updates the external view
5. Broker watches external view and updates the routing to not query the servers

Between step 3 to 5, broker will still query the server for the deleted segment, but the segment is already offloaded from the server.

# Solution
Broker should also watch on ideal state change, and update the routing when the segment is deleted or moved in the ideal state. If a segment/instance is in the external view but not in the ideal state, broker should not query the segment/instance as it will be dropped soon in the external view.