PinotConfigProvider interface (#7984)
Adding a PinotConfigProvider interface, to get TableConfig and Schema, and allow listeners to register and get notified of changes.
We already have an implementation of this in the form of TableCache in pinot-common. However, this impl cannot be used as is in our plugins, since plugins do not take the dependency on pinot-common. Accessing configs is a very common operation in plugins (e.g. minion plugins do this frequently).
Adding this interface will help us use this in our plugins and keep the dependency clean.

Adding the Listeners to the ConfigProvider, because

we can eliminate the need for restart/API invocation in case of certain operations like reload.
it's in general a good API to have on a ConfigProvider