Add consumption rate limiter for LLConsumer (#6291)
Having a consumption rate limiter for a table is beneficial in two ways:

1) Currently offline tables can be configured to have storage quota. The storage quota for tables safegaurds the pinot cluster hosting those tables in case of excessive uploads of offline segments so that the cluster won't be under provisioned. Consumption rate limiter can play similar role for hybrid tables and realtime-only tables as it limits the amount of data being ingested into the realtime tables.

2) Bursty pattern of incoming Kafka events introduces long time GC pauses on Pinot server. This behavior has been observed in some tables in LinkedIn. It's also been reported in this issue: Real time server query performance hit when catching up Kafka backlogs #6138.

This PR introduces rate limit on Kafka consumption which can be configured in StreamConfig. The consumption rate limit parameter is defined for the whole topic and the rate limit for each partition will be simply topic rate limit divided by the partition count of that topic.