Disable recursion in PinotFS copy (#8162)
* Disable recursion in PinotFS copy

All uses of PinotFS copy API involve copying a tarred segment and
untarring it. So, copying a directory recursively will not work (the
untar will fail). It also results in wastage of effort in copying
across file systems.

Also disabled the file scheme in during segment upload on the controller,
since the URL based upload is meant to provide an external URL to be
picked up by the controller.

* Removed redundant comment

* Fix to add a different API for recursive copy

* Fix lint errors