Optimize geometry serializer usage when literal is available (#8167)
Cache the Literal converted Geometry instead of computing it each time the Geometry function is called. 

This 
- save memory for processing literal transform function (previously it was done via Array.fill)
- save compute from Geometry function

In general this should apply to any chained call to LiteralTransformFunction with block --> it should be considered as a fake array instead fo a real, replicated one. This is out of scope of this PR but something to keep in mind.