In TableCache, invoke callback on config change listener during registration (#8302)
Currently when the config change listener is registered, the callback is not triggered automatically. It relies on the caller to set the current configs into the change listener, which is hard to use because there might be other changes invoking the `onChange()` callback at the same time.
This PR changes the listener registration method to directly call the `onChange()` with the current configs.