Group by order by interning (#8376)
This caches raw values extracted from dictionaries in ND group-by/order-by queries where N > 1 for low cardinality columns. This is an effective optimization because if the cardinality of dimension 1 is x and the cardinality of dimension 2 is y then the dictionary ids of dimension 1 will be decoded up to y times, and those of dimension 2 up to x times. The size of the cache is limited to 10k elements per dimension per group by query, which is similar to what gets allocated elsewhere in the query engine and does not scale with the number of segments processed.