Reduce the size of arrays allocated in DataBlockCache
We are seeing ~50% of memory allocated in a given period from these arrays.
Our use case is such that we have heavy filtering from a given segment, producing
fewer than 100 docs in the vast majority of cases.

This patch reduces the size of these arrays to be match the `_length` of each
data block. Given that we already have that length and each iteration over a segment
will likely run many times for up to 10k docs, followed by some remaining smaller
block, this should maintain the same properties of allocating arrays of 10k size,
in the case of full table scans or similar workload where there are many blocks
accessed.

However, it vastly reduces the memory churn on cases that fetch only a few rows.