Set 'shutdownInProgress' in server config when server starts before the startup check (#8525)
For new joined server, or server not shut down normal previously, the shutdownInProgress won't be set properly when it starts. This can cause brokers routing queries to it before it finishes the startup check and cause performance issue.
This PR adds the flag if it is not set when server starts. Also reduces the updates of instance config by checking before setting.