Use MultiThreadedHttpConnectionManager in SegmentStatusChecker
Currently, SegmentStatusChecker initializes an object of type
SimpleHttpConnectionManager. Further down the stack, we use this object to
issue parallel HTTP GET requests to multiple servers containing segments for
the table (in MultiGetRequest). However, SimpleHttpConnectionManager is not
thread safe. So we should be using an instance of MultithreadedHttpConnectionManager.

Added a TODO to improve MultiGetRequest.java to only work on HttpConnectionManager
object that is an instanceOf MultithreadedHttpConnectionManager.