Store batch message ids vs calculating them (#8612)
Currently, Pulsar batch message ids to start the consumer are calculated by assuming each batch is of exact BATCH_SIZE.
However, in pulsar producer configuration, that is the max size and sometimes the actual batches can be smaller than the particular size.
The only way is to keep track of all batch ids in a map. This will help reduce the instances of flakiness in the future.