Scalable Gapfill Implementation for Avg/Count/Sum (#8647)
This is the performance optimization for avg/count/sum aggregation of gap filled result. These aggregation functions are about counting. It is time-consuming to create the gapfilled rows when doing the aggregation. This PR will not create the gap filled rows when doing the counting. This will give 10x performance improvement.