[BugFix] Avoid reporting negative values for server latency.
While working on testing changes in #8808, noticed that we report negative
values for ResponseDelayMs in high QPS usecases. This can happen if the response
from the server is received before the query is marked as submitted. This
typically happens if request is sent to server but there's a delay in calling the
callback in _channel.writeAndFlush(cb).