Add controller API for reload segment task status (#8828)
This PR adds a controller API that can be used to get the current status of a 'reload segments' task. When a segment reload task is submitted, an identifier will be returned, and the task details will be stored into ZK. Users can then call the newly added API with the task ID, to get the status details of the reload task.
Another API has been added to get all historical reload task ids.