[multisage] fix queries with LIMIT and no ORDER BY (#9498)
before this change, the optimizer would allow pushing down
of LIMIT clauses without any ORDER BY. This is a problem,
because it doesn't have any top-level LIMIT which would
limit the final results of the query. This PR fixes that
problem, and I will follow this up with another PR that
adds an additional SORT stage at the top level.