bubble up error messages from broker (#9754)
This PR improves error message handling on the broker and ensures that root causes are returned to the user. There's probably a lot of additional work to be done to make sure that the response is improved for all different types of error messages, but this is an improvement over the status quo.