UI: show segment debug details when segment is in bad state (#9700)
What does this PR do?
show segment debug details when segment is in bad state
Issue:
One of the recurring issue is that if a segments show as BAD status, users have no idea why.
To get more info about error state then we need to check the debug API or check the server log for the segment.
We need to have some way in UI so that users can get immediate feedback on why the segment is in BAD state
Description:
Show a help icon besides Bad state. This also helps grab user attention when Bad state is visible
On click of help icon a dialog will open with segment level debug details.