BugFix: Handle BYTES column type for partitioning. (#10110)
The MurmurPartition function does not work with byte[] data type as it converts the value to String `(value.toString().getBytes(UTF_8)` first.
In the offline flow, we already convert byte[] type to ByteArray for the MurmurPartition function to handle it correctly. This PR addresses the real-time ingestion path to also do the same conversion to ByteArray.