[feature] Add flags to indicate if  FST, H3, and Text indexes are on a segment (#10835)
This PR augments the Segment Metadata returned by the Controller REST API /segments/{tableName}/{segmentName}/metadata with flags indicating if columns in a segment have any of the following indexes: H3, FST, and Text. Currently, Pinot has flags for every index but these 3. Without this change, it is not possible to tell if these 3 indexes are on columns in a segment through the metadata returned by the Pinot Controller.