Make org.apache.pinot.client.BrokerResponse public
The org.apache.pinot.client.PinotClientTransport class is an
interface to implement the client transport for Pinot Java
Client connections. Currently, there is a default implementation
provided by Pinot Java client which uses AsyncHttpClient.

We want to implement the PinotClientTransport interface to
provide a Vert.x adapter for the Pinot Java Client but currently
it is not possible because the interface methods return
BrokerResponse class which is package-private. Hence, making
BrokerResponse public so that the interface can be implemented
externally.

Partially fixes #10924.