Add fromController method that accepts a PinotClientTransport
`ConnectionFactory` class is used to create instances on pinot client's
`Connection` class. There are different methods available depending upon
the kind of broker that we want to use. All of these except `fromController`
also provide a variant to accept a `PinotClientTransport` from the user.

We are wrapping around the `ConnectionFactory` class in the Vert.x wrapper
for pinot client to provide our custom transport. Since the constructors
of the `Connection` class are package-private, we cannot directly add this
method to an external package. Further, this method seems to be absent from
the class only as a mistake (there are methods to specify custom transports
with all other brokers).

Hence, adding this method.

Partially fixes #10924.