Make all accessor methods on BrokerResponse public
In #10932, I requested for BrokerResponse class to be made public so that an
external implementation of PinotClientTransport could be created. At that time,
I failed to rectify that most of the methods to access data from BrokerResponse
class are package-private which again hinders using them in a custom transport.
Hence, opening this PR to make those public methods too.

CC: @tsegismont