Add UltraLogLog support
UltraLogLog is a variant of HyperLogLog from dynatrace, with an implementation available in hash4j under the apache license.

This adds support for using it in the ways you'd use HLL in Pinot.

When using it against normal java types, wyhash 4 is used as the default hashing algorithm, when bringing your own serialized sketches you can use any.

* supports `UltraLogLog` as a data type
  * the serialization format includes serializing the P value as the first byte in the data stream, so the width is known in streams
* adds `DistinctCountULL` and `DistinctCountRawULL` for use in SQL
  * raw outputs Base64 encoded bytes that can be fed into `UltraLogLog.wrap` in other services
* adds startree support
* adds merge rollup support
* new transformaction functions added
  * `toULL` allows turning data into a ULL
  * `fromULL` lets you import ULLs encoded as byte arrays outside pinot