Theta Sketch Aggregation Enhancements
Introduces additional parameters to the DistinctCountThetaSketch aggregation function that
give the end-user more control over how sketches are merged.  The defaults are selected
to ensure that the behaviour remains unchanged over the current implementation.

Furthermore, an accumulator custom object is added to ensure that pairwise union
operations are avoided as much as possible.  Instead, sketches can be aggregated
and merged when a threshold is met.