Theta Sketch Aggregation Enhancements (#12042)
* Theta Sketch Aggregation Enhancements

Introduces additional parameters to the DistinctCountThetaSketch aggregation function that
give the end-user more control over how sketches are merged.  The defaults are selected
to ensure that the behaviour remains unchanged over the current implementation.

Furthermore, an accumulator custom object is added to ensure that pairwise union
operations are avoided as much as possible.  Instead, sketches can be aggregated
and merged when a threshold is met.

* Use correct naming convention for private variable

* Fetch flaky test edge-case

* Decrease default constant value

This better aligns to the default nominal values parameter
that is used in the query aggregation function.

* Attempt 2: Simplify implementation

Removes intermediate array list to buffer/accumulate sketch
elements.  Instead, inputs are fed directly to the underlying
union.  This ensures that the memory usage of the merge is
kept under control.

* Rename sampling probability parameter

* Minor code improvements

* Revert "Attempt 2: Simplify implementation"

This reverts commit 2ed38f395ae1e3071637f58da2386e867dbc80e3.

* Add toString methods for supported sketches to enable debugging

* Additional inline commentary on early stop optimization.

* Refactor serializer for Theta to remove temp variables