Refactor logger initialization in multiple classes
This commit addresses a widespread issue of Logger instances being incorrectly initialized in multiple classes. The Logger instances were mostly being created with the wrong class as argument to the LoggerFactory.getLogger() method. This issue has now been fixed by replacing the incorrect class names with the corresponding class in which the Logger instance is initialized. #12127