Implement IN and NOT_IN filters in FilterOperand class (#12285 fix)
Added IN and NOT_IN cases to the FilterOperand class and TransformOperandFactory classes. The PR is related to the 12285 bug:  using of the IN expression in the CASE statement within the projection part of the SELECT multistage query. When this expression is used, the query execution fails with an IllegalStateException error, indicating that the function IN cannot be found.