Fix query response time metric
As of PR #13179, the query response time metrics started to get extremely high for
all tables. The root cause tured out to be a missed initialization in which
the query arrival time was set to 0 by default.

The refactor broke the contract that the broker request handler always returns the
time taken to handle the request.

Added code to initialize the arrival time.

Since the refactor attempted to include client pre-processing time, added this code
under the condition that the value is not already set.

Ideally, callers of the method should add on their own pre and post-processing time
to the time returned by the handleRequest() method.