Support of classrealms and new plugin structure (#13930)
With classrealms there will be much more control over the classloaders.
Next to the shaded plugins, a plugin-directory may contain a pinot-plugin.properties (to switch from PluginClassloader to ClassRealms).

if there is no pinot-plugin.properties, the original behavior is used.
if there is a pinot-plugin.properties, a classrealm will be created: every direct file or directory in the plugin-directory will be added to the classrealm.

It is possible to let shaded plugins use a classrealm. In that case the classes in the shaded jar will be loaded before the system classloader (as done by the PluginLoader).

this implementation doesn't unpack plugin-zips as created with the assembly descriptor. Instead is assumes it is unpacked before pinot starts.