Add support for configuring Theta and Tuple aggregation function params in the ST index
This patch introduces a mechanism to allow configuring the aggregation function parameters
for a star-tree index for Tuple and Theta sketches.  Any existing aggregation that has a
precision greater or equal to that of the query precision is selected as a candidate.
The behaviour of the CPC aggregator has been changed accordingly.