Improve reported dataTypes on V1 when no rows returned (#14836)
V1 just uses STRING as default data type for all columns when no data has been returned by servers.

This is an attempt to improve that by different strategies, being the preferred one use the schema validation implemented on V2 and reuse the data types determined by it. As a failback and only suitable for non-transformed fields, the base table schema is used. Otherwise, the default STRING type is still returned.