FIX: Modifying the helm chart used for Pinot Quickstart (#14952)
This bugfix fixes the issue with #14696

According to the documentation for installation of Pinot in k8s using helm charts (https://docs.pinot.apache.org/basics/getting-started/kubernetes-quickstart#start-pinot-with-helm), the namespace is created manually. This creates a conflict with the namespace.yaml present in the templates, leading to failure of deployment.

Signed-off-by: Subhashini Narayanaswamy  <subhashini.narayanaswamy@mercedes-benz.com>